Harden the support, VM, and IL layers against malformed input
A cross-cutting defensive-robustness pass over the foundation libraries
(src/support), the bytecode/tree-walking VM, the IL pipeline, and the
runtime heap. The common thread is replacing silent failure and
undefined behavior with explicit, fail-fast guards: asserts become
observable traps, unchecked parses become range-checked ones, raw
new[]/delete[] become allocator-driven construct/destroy, and interior
pointers are validated against live allocations before they are
dereferenced. New unit tests pin the added contracts.

Support library (src/support):

- arena: guard every pointer/size addition in tryAllocate against
  uintptr_t/size_t overflow, fix the returned pointer to data + aligned
  offset (was data + offset - size), throw bad_alloc on bad alignment
  instead of returning null, and reserve the destructor slot in create<T>
  before constructing so a failed push_back cannot orphan a live object.

- small_vector: switch inline storage from a default-constructed T[N] to
  raw alignas(T) bytes and drive all element lifetime through the
  allocator with construct_at / destroy_at. This makes the container
  correct for non-trivial and non-default-constructible element types,
  with exception-safe copy/move/reserve/emplace_back/resize and checked
  geometric growth against max_size.

- source_manager: serialize files_/path_to_id_/lineCache_ behind a mutex;
  track a parallel disk_paths_ of absolute, lexically-normalized paths so
  lazy line loads are stable across cwd changes; add hasLine (distinguish
  "no such line" from "empty line"), invalidateSource, and a setSource
  that validates the file id. addFile reserves before insert and defers
  the id bump until every mutation succeeds.

- string_interner: make move-construct/move-assign explicit and
  exception-safe by building the replacement string_view map from the
  moved storage before swapping it in (keys can't dangle), pop storage on
  duplicate/throw in intern, and add contains() and lookupOptional() so
  callers can tell "absent" from "empty string".

- diagnostics: validate UTF-8 sequences and emit U+FFFD for malformed
  bytes instead of escaping each byte; tighten machine-readable ranges
  (isValid -> isConcrete, zero-width ranges rejected, new isInsertion);
  guard the source-line and path emitters; normalize legacy captured
  messages (strip trailing CR/LF and severity prefixes) with a fallback
  message when capture is empty.

VM (src/vm):

- Memory access classification: new VM::MemoryAccessInfo plus
  classifyMemoryAccess validates every load/store/gep range against the
  frame stack, mutable globals (sizes now tracked in ProgramState), or
  the runtime heap before dereferencing, taking mutableGlobalMutex when
  the range is a shared global. Missing-operand instructions trap as
  malformed rather than reading past the operand vector.

- Indirect calls validate the callee against fnMap before dispatch;
  fast-path runtime calls switch from ">= N" to strict "!= N" arity and
  always trap rather than silently falling through.

- Overflow/range hardening: operand-cache offsets widen to size_t with an
  explicit overflow check; switch.i32 metadata is fully validated
  (default label, scrutinee, case-count parity, i32-range case values)
  before the dispatch cache is built; fptosi uses an exclusive 2^63 bound
  and "< 0.0" in place of signbit.

- Branch/switch terminators replace asserts with traps for invalid or
  out-of-range targets and branch-argument count mismatches; every trap
  path now returns an ExecResult marked returned so dispatch stops.

- Global initializers are parsed with range-checked integer/float helpers
  (errno + full-consume + bounds) so out-of-range literals are rejected
  instead of wrapping; initializer storage is owned by a unique_ptr that
  cleans up if the size-map insert throws.

- Runtime bridge gains callMutable so by-reference out-args are copied
  back via the signature's ownedOutArgMask; mapKind now maps string
  params to Str (not Ptr); validateArgumentCount returns bool and traps.

- main's return value is validated (integer or void only; string results
  released) before becoming the process exit code; runFunctionLoop is
  wrapped so frame buffers are released on exception; per-instruction
  polling folds a POSIX signal-safe sig_atomic_t interrupt flag into the
  atomic epoch.

- String register slots are released before being overwritten or zeroed,
  and TCO defers all frame mutation until after every non-mutating
  eligibility check passes.

IL pipeline (src/il):

- Fail-fast identifier validation: the serializer, instruction parser,
  and extern/function/global/block verifiers reject empty or
  syntactically invalid IL identifiers via isValidILIdentifier /
  checkedIdentifier instead of emitting or accepting malformed text;
  CFG construction throws on an unknown branch label rather than dropping
  the edge.

- Float/bit correctness: FPCast and ValueKey use memcpy for bit
  reinterpretation (no inactive-union-member UB); the literal parser
  rejects misplaced underscore separators and parses floats through a
  classic-locale istringstream for deterministic, trailing-garbage-free
  results.

- Loop passes (rotate/unroll/simplify) recompute LoopInfo/CFG after each
  transform instead of trusting a stale snapshot, and tighten
  latch/entry-edge bookkeeping by instruction index.

- Pass management: PassRegistry fingerprints the module around legacy
  void passes to report accurate PreservedAnalyses, and schedules the
  whole-module analysis consumers (gvn, indvars, licm, earlycse,
  loop-rotate/simplify, check-opt) sequentially; PipelineExecutor skips
  cleanup passes already run since the last change and aggregates parallel
  worker analysis counts; AnalysisManager evicts cache entries for
  removed/changed functions and gains targeted invalidateFunctionResult.

- EarlyCSE/GVN replace uses through the new dominance-aware
  replaceUsesDominatedBy to avoid use-before-def when the replacement is a
  temporary; nextTempId checks for id-space exhaustion; DSE erases
  instruction sites deterministically and treats Trap/TrapFromErr as
  function exits.

- ModuleLinker now intersects (rather than unions) extern pure/readonly/
  nothrow attributes so a merged extern only advertises guarantees every
  declaration makes.

Runtime (src/runtime):

- rt_heap gains rt_heap_contains_range, which scans the live-allocation
  registry under the heap lock and accepts an interior pointer only when
  the whole [ptr, ptr+bytes) range lies inside a single live payload
  (magic/refcount/size validated, boundary crossings and overflows
  rejected). The VM memory-access path uses it; the symbol is exposed via
  RuntimeSurfacePolicy.

Tests:

- test_support adds ~200 lines covering SmallVector non-trivial element
  lifetimes (inline->heap spill, move, pop/clear destructor counts,
  no-default-constructor types), StringInterner move semantics and
  contains/lookupOptional, SourceRange isConcrete/isInsertion, blank-line
  and unknown-file/invalid-UTF-8/unknown-severity diagnostics, lazy
  line-cache invalidation with CRLF normalization, relative-path capture,
  and DiagCapture pre-formatted/empty handling.

- ModuleLinkerTests asserts attributes are intersected (readonly/nothrow
  false) and is renamed accordingly; CastOpsTests treats -0.0 as a valid
  fptoui input returning 0 rather than a trap; the BASIC verify-diag
  fixture uses the unprefixed name "main".

Docs:

- Refresh the v0.2.7 release-notes draft already staged in the worktree.

Validation:

- git diff --check

+++ src/vm/OpHandlerUtils.hpp
@@ -360,7 +360,7 @@ struct OperandDispatcher {
         const auto *state = detail::VMAccess::currentExecState(vm);
         const auto *bc = state ? state->blockCache : nullptr;
         if (bc && state->ip < bc->instrOpOffset.size()) [[likely]] {
-            const uint32_t off = bc->instrOpOffset[state->ip];
+            const size_t off = bc->instrOpOffset[state->ip];
             lhs = evalFast(vm, fr, bc->resolvedOps[off], in.operands[0]);
             rhs = evalFast(vm, fr, bc->resolvedOps[off + 1], in.operands[1]);
         } else {
@@ -383,7 +383,7 @@ struct OperandDispatcher {
         const auto *state = detail::VMAccess::currentExecState(vm);
         const auto *bc = state ? state->blockCache : nullptr;
         if (bc && state->ip < bc->instrOpOffset.size()) [[likely]] {
-            const uint32_t off = bc->instrOpOffset[state->ip];
+            const size_t off = bc->instrOpOffset[state->ip];
             lhs = evalFast(vm, fr, bc->resolvedOps[off], in.operands[0]);
             rhs = evalFast(vm, fr, bc->resolvedOps[off + 1], in.operands[1]);
         } else {
